+++ hub3/ead/parser.go
@@ -19,25 +19,16 @@ import (
 	"context"
 	"encoding/xml"
 	"fmt"
-	"io"
 	"io/ioutil"
 	"log"
-	"net/http"
-	"os"
 	"strings"
 	"sync/atomic"
-	"time"
 	"unicode"
 
-	"github.com/delving/hub3/config"
 	"github.com/delving/hub3/hub3/fragments"
-	"github.com/delving/hub3/hub3/models"
 	"github.com/delving/hub3/ikuzo/domain/domainpb"
-	"github.com/go-chi/render"
 	r "github.com/kiivihal/rdf2go"
 	"github.com/microcosm-cc/bluemonday"
-	"github.com/pkg/errors"
-	"github.com/src-d/go-git/plumbing"
 )
 
 type BulkIndex interface {
@@ -74,200 +65,6 @@ func eadParse(src []byte) (*Cead, error) {
 	ead := new(Cead)
 	err := xml.Unmarshal(src, ead)
 	return ead, err
-}
-
-func ProcessEAD(r io.Reader, headerSize int64, spec string, bi BulkIndex, tags ...string) (*NodeConfig, error) {
-	os.MkdirAll(config.Config.EAD.CacheDir, os.ModePerm)
-
-	f, err := ioutil.TempFile(config.Config.EAD.CacheDir, "*")
-	if err != nil {
-		log.Printf("Unable to create output file %s; %s", spec, err)
-		return nil, err
-	}
-	defer f.Close()
-
-	buf := bytes.NewBuffer(make([]byte, 0, headerSize))
-	_, err = io.Copy(f, io.TeeReader(r, buf))
-
-	if err != nil {
-		return nil, err
-	}
-
-	cead, err := eadParse(buf.Bytes())
-	if err != nil {
-		log.Printf("Error during parsing; %s", err)
-		return nil, err
-	}
-
-	if spec == "" {
-		spec = cead.Ceadheader.Ceadid.EadID
-		if strings.Contains(spec, "/") {
-			spec = strings.ReplaceAll(spec, "/", ".")
-		}
-	}
-
-	f.Close()
-
-	ds, _, err := models.GetOrCreateDataSet(spec)
-	if err != nil {
-		log.Printf("Unable to get DataSet for %s\n", spec)
-		return nil, err
-	}
-	// check for cache entry
-	hash := plumbing.ComputeHash(plumbing.BlobObject, buf.Bytes())
-	ds.Fingerprint = hash.String()
-
-	basePath := GetDataPath(spec)
-	os.MkdirAll(basePath, os.ModePerm)
-	os.Rename(f.Name(), fmt.Sprintf("%s/%s.xml", basePath, spec))
-
-	ds, err = ds.IncrementRevision()
-	if err != nil {
-		log.Printf("Unable to increment %s\n", spec)
-		return nil, err
-	}
-
-	// set basics for ead
-	ds.Label = cead.Ceadheader.GetTitle()
-	ds.Period = cead.Carchdesc.GetPeriods()
-
-	// description must be set to empty
-	ds.Description = ""
-
-	cfg := NewNodeConfig(context.Background())
-	cfg.CreateTree = CreateTree
-	cfg.Spec = spec
-	cfg.OrgID = config.Config.OrgID
-	cfg.Revision = int32(ds.Revision)
-	cfg.Tags = tags
-
-	// create desciption
-	desc, err := NewDescription(cead)
-	if err != nil {
-		return nil, errors.Wrapf(err, "Unable to create description")
-	}
-
-	cfg.Title = []string{desc.Summary.File.Title}
-	cfg.TitleShort = desc.Summary.FindingAid.ShortTitle
-
-	descIndex := NewDescriptionIndex(spec)
-	err = descIndex.CreateFrom(desc)
-	if err != nil {
-		return nil, fmt.Errorf("unable to create DescriptionIndex; %w", err)
-	}
-
-	err = descIndex.Write()
-	if err != nil {
-		return nil, fmt.Errorf("unable to write DescriptionIndex; %w", err)
-	}
-
-	err = desc.Write()
-	if err != nil {
-		return nil, fmt.Errorf("unable to write description; %w", err)
-	}
-
-	// save description
-	var unitInfo *UnitInfo
-	if desc.Summary.FindingAid != nil && desc.Summary.FindingAid.UnitInfo != nil {
-		unitInfo = desc.Summary.FindingAid.UnitInfo
-		ds.Length = unitInfo.Length
-		ds.Files = unitInfo.Files
-		ds.Abstract = unitInfo.Abstract
-		ds.Language = unitInfo.Language
-		ds.Material = unitInfo.Material
-		ds.ArchiveCreator = unitInfo.Origin
-	}
-
-	nl, _, err := cead.Carchdesc.Cdsc.NewNodeList(cfg)
-	if err != nil {
-		log.Printf("Error during parsing; %s", err)
-		return cfg, err
-	}
-
-	ds.MetsFiles = int(cfg.MetsCounter.GetCount())
-	ds.Clevels = int(cfg.Counter.GetCount())
-
-	err = ds.Save()
-	if err != nil {
-		return nil, errors.Wrapf(err, "Unable to save dataset")
-	}
-
-	err = cead.SaveDescription(cfg, unitInfo, bi)
-	if err != nil {
-		log.Printf("Unable to save description for %s; %#v", spec, err)
-		return nil, errors.Wrapf(err, "Unable to create index representation of the description")
-	}
-
-	// write error log
-	if len(cfg.Errors) != 0 {
-		errs, err := cfg.ErrorToCSV()
-		if err != nil {
-			log.Printf("unable to get error csv: %#v", err)
-			return nil, err
-		}
-
-		err = ioutil.WriteFile(
-			fmt.Sprintf("%s/errors.csv", basePath),
-			errs,
-			0644,
-		)
-		if err != nil {
-			return nil, errors.Wrapf(err, "Unable to EAD erros to disk")
-		}
-
-	}
-
-	if bi != nil {
-		start := time.Now()
-		err := nl.ESSave(cfg, bi)
-		if err != nil {
-			log.Printf("Unable to save nodes; %s", err)
-		}
-
-		// TODO(kiivihal): decide what to do with drop orphans later
-		// _, err = ds.DropOrphans(context.TODO(), bi, nil)
-		// if err != nil {
-		// log.Printf("Unable to drop orphans; %s", err)
-		// }
-		end := time.Since(start)
-		log.Printf("saving %s with %d records took: %s", spec, cfg.Counter.GetCount(), end)
-	}
-
-	return cfg, nil
-}
-
-func ProcessUpload(r *http.Request, w http.ResponseWriter, spec string, bi BulkIndex) (uint64, error) {
-	in, header, err := r.FormFile("ead")
-	if err != nil {
-		return uint64(0), err
-	}
-
-	tags := r.FormValue("tags")
-
-	defer in.Close()
-	defer func() {
-		err = r.MultipartForm.RemoveAll()
-	}()
-
-	cfg, err := ProcessEAD(in, header.Size, spec, bi, strings.Split(tags, ",")...)
-	if err != nil {
-		//http.Error(w, err.Error(), http.StatusInternalServerError)
-		return 0, err
-	}
-
-	render.PlainText(w, r, fmt.Sprintf("Processed %d for dataset %s\n", cfg.Counter.GetCount(), spec))
-	log.Printf("nr of errors: %d", len(cfg.Errors))
-	if len(cfg.Errors) > 0 {
-		render.PlainText(w, r, fmt.Sprintf("Duplicate inventory numbers %d for dataset %s\n", len(cfg.Errors), spec))
-		d, err := cfg.ErrorToCSV()
-		if err != nil {
-			return uint64(0), err
-		}
-		w.Write(d)
-	}
-
-	return cfg.Counter.GetCount(), nil
-
 }
 
 type CLevel interface {